pipelines/execute_bq_query/pyproject.toml
[project]
name = "namaph-council-pipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "db-dtypes>=1.4.0",
    "functions-framework>=3.8.2",
    "google-cloud-bigquery>=3.29.0",
    "google-cloud-storage>=2.19.0",
    "pandas>=2.2.3",
    "requests>=2.32.3",
]
